Removed as many static references as makes sense
For the most part I think the api is fairly clean and I'm having an issue completely ripping out all the things that might actually end up needing to remain static.

Just a few things I think that might need renamed (WebDriverHolder) and made package private to better protect some things that are core.  Protecting the setWebDriver method seems critical to me as someone might do something to null that out and close it manually which would then bubble out as an NPE when the framework attempted to tear it down.