Updated log even writes
I left the screenshot piece as a placeholder as I do not intend to continue showing screenshots inline, especially now that we may have multiple images per method. I plan to ensure the report is at least writing correctly, then I'll come back and flesh out the screenshot viewer, which will be more than just embedding an image (so long as I wish to have a link open a modal to show the image).